backend/pyproject.toml
[project]
name = "insync"
version = "0.1.0"
description = "YouTube/Spotify to Jellyfin playlist sync service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.15.2",
  "apscheduler>=3.11.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "mutagen>=1.47.0",
  "pydantic-settings>=2.10.1",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.20",
  "spotipy>=2.25.1",
  "sqlalchemy>=2.0.41",
  "uvicorn[standard]>=0.35.0",
  "yt-dlp>=2025.6.30",
  "yt-dlp-ejs>=0.8.0",
  "ytmusicapi>=1.11.1",
]

[dependency-groups]
dev = [
  "pytest>=8.4.1",
  "ruff>=0.12.7",
  "ty>=0.0.1a17",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py313"
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "SIM"]

[tool.ty.environment]
python = "./.venv"

[tool.ty.src]
include = ["app"]
